riscv: add meta.arch to make it easier to re-use
the meta.cpu used to be "rv32imac" which was fine as an arch but I split
it to more finegrained to better support different vendor cpus.

Probably still want something like arch though, this lets you pull out
the rv32imac tag for all the riscv cpus, regardless of type.

Signed-off-by: Karl Palsson <karlp@tweak.net.au>