build: support variant_dir more.
This is not 100%.  We've removed dependencies on the linker scripts,
though it seems to rebuild anyway when I edit them.

Not sure this is final, or cleanest, but this was laks side to get
multiple variants building.

Tested with a "plain" laks build as well, which then uses the LIBPATH to
find the plain old generated.ld